requirements.txt
PyQt6>=6.6.0
Pillow>=10.0.0
psd-tools>=1.9.0
numpy>=1.24.0
